Improving deposit bonus logic
