+++ server/src/tasks/sqs/funder/token.js
@@ -1,9 +1,8 @@
-const mongoose = require('mongoose')
 const { createNetwork } = require('@utils/web3')
 const { transfer } = require('@utils/token')
-const QueueJob = mongoose.model('QueueJob')
 const { fetchToken, adjustDecimals } = require('@utils/token')
 const { notifyReceiver } = require('@services/firebase')
+const { validateFundingLimitPerUser } = require('@utils/jobs')
 
 const fundToken = async (account, { phoneNumber, receiverAddress, identifier, tokenAddress, communityAddress, bonusType, bonusMaxTimesLimit, bonusAmount }, job) => {
   const network = createNetwork('home', account)
@@ -15,19 +14,7 @@ const fundToken = async (account, { phoneNumber, receiverAddress, identifier, to
   if (!identifier) {
     throw Error(`No identifier defined. [phoneNumber: ${phoneNumber}, receiverAddress: ${receiverAddress}, tokenAddress: ${tokenAddress}, communityAddress: ${communityAddress}, bonusType: ${bonusType}]`)
   }
-
-  const fundingsCountForPhoneNumber = await QueueJob.find({
-    name: 'fundToken',
-    status: { $ne: 'failed' },
-    _id: { $ne: job._id },
-    'data.phoneNumber': phoneNumber,
-    'data.tokenAddress': tokenAddress,
-    'data.communityAddress': communityAddress,
-    'data.receiverAddress': receiverAddress,
-    'data.bonusType': bonusType
-  }).countDocuments()
-
-  if (fundingsCountForPhoneNumber >= bonusMaxTimesLimit) {
+  if (!validateFundingLimitPerUser({ job, phoneNumber, tokenAddress, communityAddress, receiverAddress, bonusType, bonusMaxTimesLimit })) {
     throw Error(`Join bonus reached maximum times ${bonusMaxTimesLimit}. [phoneNumber: ${phoneNumber}, receiverAddress: ${receiverAddress}, tokenAddress: ${tokenAddress}, communityAddress: ${communityAddress}, bonusType: ${bonusType}]`)
   }
   const amountInWei = adjustDecimals(bonusAmount, 0, decimals)
